requirements.txt
Python==3.10.11

Django==5.1.1

pillow==10.4.0

django-summernote==0.8.20.0

beautifulsoup4==4.12.3

python-dotenv==1.0.1

google-cloud-aiplatform==1.73.0

openai==1.54.4

1